requirements.txt
boto3==1.34.152
requests==2.32.3
PyYAML==6.0.1
